Refactor lib/galaxy/objectstore/test_helpers.py to package-less objectstore_helpers.py.
Need more symlinks and such now.
