+++ test/unit/tools/test_history_imp_exp.py
@@ -7,10 +7,10 @@ from tempfile import mkdtemp
 
 from galaxy import model
 from galaxy.exceptions import MalformedContents
-from galaxy.objectstore.test_helpers import TestConfig
 from galaxy.tools.imp_exp import JobExportHistoryArchiveWrapper, JobImportHistoryArchiveWrapper, unpack_tar_gz_archive
 from galaxy.tools.imp_exp.export_history import create_archive
 from ..unittest_utils.galaxy_mock import MockApp
+from ..unittest_utils.objectstore_helpers import TestConfig
 
 
 # good enough for the very specific tests we're writing as of now...
